+++ LMSAdmin/src/main/java/com/bridgelabz/lms/service/CandidateServiceImpl.java
@@ -3,16 +3,14 @@ package com.bridgelabz.lms.service;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import com.bridgelabz.lms.dto.UpdateDto;
-import com.bridgelabz.lms.dto.UserDTO;
+import com.bridgelabz.lms.dto.CandidateHiredDTO;
+import com.bridgelabz.lms.dto.UpdateHiringDto;
 import com.bridgelabz.lms.exception.CandidateRegistrationException;
-import com.bridgelabz.lms.exception.InvalidDetailsException;
 import com.bridgelabz.lms.model.Candidate;
 import com.bridgelabz.lms.repository.CandidateRepository;
 import com.bridgelabz.lms.response.Response;
@@ -38,7 +36,7 @@ public class CandidateServiceImpl implements CandidateService {
 	private CandidateRepository userrepo;
 
 	@Override
-	public Response registerCandidate(UserDTO dto) throws InvalidDetailsException {
+	public Response registerCandidate(CandidateHiredDTO dto) {
 		Optional<Candidate> isuserprsent = userrepo.isEmailExists(dto.getEmail());
 		if (isuserprsent.isPresent()) {
 			throw new CandidateRegistrationException("invalid details", null, 400, "true");
@@ -48,7 +46,7 @@ public class CandidateServiceImpl implements CandidateService {
 			user.setUpdateStamp(LocalDateTime.now());
 			userrepo.save(user);
 			log.info(user.getFirstName() + " registered " + "date:" + user.getCreatorStamp());
-			String body = "http://localhost:8080/verifyemail/" + jwt.jwtToken(user.getId());
+			String body = "http://localhost:8081/verifyemail/" + jwt.jwtToken(user.getId());
 			System.out.println(body);
 			jms.sendEmail(user.getEmail(), "verification email", body);
 			return new Response("regitration sucess", user, 201, "true");
@@ -56,37 +54,34 @@ public class CandidateServiceImpl implements CandidateService {
 
 	}
 
-	@Override
-	public Candidate verify(String token) {
-		long id = jwt.parseJWT(token);
-		Candidate user = userrepo.isIdExists(id)
-				.orElseThrow(() -> new CandidateRegistrationException("user not exists", HttpStatus.OK, id, "false"));
-		user.setVerifyEmail(true);
-		userrepo.save(user);
-		return user;
-	}
-
+	
 	@Override
 	public Response getAllHiredcandidates() {
 		List<Candidate> isUserPresent = userrepo.findAll();
 		return new Response("List of HiredCandidates are", isUserPresent, 200, "true");
 	}
 
-	public Candidate getUserById(String token) {
-		long id = jwt.parseJWT(token);
-		return userrepo.getUserById(id)
-				.orElseThrow(() -> new CandidateRegistrationException("user not exists", HttpStatus.OK, id, "false"));
+
+	@Override
+	public Candidate verify(String token) {
+		long id=jwt.parseJWT(token);
+		log.debug(token);
+		Candidate user=userrepo.isIdExists(id).orElseThrow(() -> new CandidateRegistrationException("user not exists",HttpStatus.OK,id,"false"));
+		user.setVerifyEmail(true);
+		userrepo.save(user);
+		return user;
 	}
+
 
 	@Override
 	public Candidate getCandidateProfileById(Long id) {
 		return userrepo.getCandidateProfileById(id)
-				.orElseThrow(() -> new CandidateRegistrationException("user not exists", HttpStatus.OK, id, "false"));
-	}
+		.orElseThrow(() -> new CandidateRegistrationException("user not exists", HttpStatus.OK, id, "false"));
+}
+
 
 	@Override
-	public Response updateCandidate(Long id, UpdateDto dto) {
-		// int Id = tokenutil.decodeToken(token);
+	public Response updateCandidate(Long id, UpdateHiringDto dto) {
 		Optional<Candidate> isUserPresent = userrepo.findById(id);
 		if (isUserPresent.isPresent()) {
 			isUserPresent.get().setFirstName(dto.getFirstName());
@@ -103,19 +98,29 @@ public class CandidateServiceImpl implements CandidateService {
 			isUserPresent.get().setPermanentAddress(dto.getPermanentAddress());
 			isUserPresent.get().setProfileImage(dto.getProfileImage());
 			isUserPresent.get().setFolderId(dto.getFolderId());
+			isUserPresent.get().setCreatorStamp(dto.getCreatorStamp());
 			// isUserPresent.get().setStatus(lmsCandidateHiring.getStatus());
 			isUserPresent.get().setUpdateStamp(LocalDateTime.now());
 			System.out.println(isUserPresent);
 			userrepo.save(isUserPresent.get());
-			return new Response("Candidate Added:", 200, "isUserPresent");
+			return new Response("regitration sucess", isUserPresent, 201, "true");
 		} else {
 			throw new CandidateRegistrationException("invalid details", null, 400, "true");
 		}
 	}
 
+
 	@Override
 	public void deleteCandidateHiringById(Long id) {
-		userrepo.deleteById(id);
+		Optional<Candidate> isUserPresent = userrepo.findById(id);
+		if (isUserPresent.isPresent()) {
+			userrepo.deleteById(id);
+		}else {
+			 new CandidateRegistrationException("Candidate to be Delete Not found",null,404,"true");
+		}
 	}
-
 }
+	
+	
+
+
